Move ftp availibility checker
