+++ src/main/java/uk/gov/hmcts/reform/sendletter/services/FtpAvailabilityChecker.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.slc.services;
+package uk.gov.hmcts.reform.sendletter.services;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
