+++ src/main/resources/application.yaml
@@ -15,6 +15,9 @@ idam:
 
 ftp:
   smoke-test-target-folder: ${FTP_SMOKE_TEST_TARGET_FOLDER:SMOKE_TEST}
+  downtime:
+    from: ${FTP_DOWNTIME_FROM:16:00}
+    to: ${FTP_DOWNTIME_TO:17:00}
 
 spring:
   application:
